+++ app/Http/Controllers/ImageController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use Dotenv\Repository\RepositoryInterface;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\File;
 use Illuminate\Support\Str;
@@ -10,55 +11,35 @@ use Illuminate\Support\Facades\Validator;
 
 class ImageController extends Controller
 {
-    public static function create($image, $to = "products"){
-        $name = time().'_'.Str::random(10).'.'.$image->getClientOriginalExtension();
-        $image->storeAs('public/images/'.$to, $name);
-        $NewName = asset('/storage/images/'.$to.'/'.$name);
-
-        return $NewName;
-    }
-
-    public static function deleteFile($fileName, $from = 'products')
-    {
-        $path = storage_path('app/public/images/'.$from.'/'.$fileName);
-        
-        return File::delete($path);
-    }
-
     public function store(Request $request)
     {
         $validation = Validator::make($request->all(), [
-            'images'=>'required|image'
+            'images'=>'required'
         ]);
 
         if($validation->fails()){
             return ResponseController::error($validation->errors()->first(), 422);
         }
 
-        $images = $request->images;
+        $images = $request->file("images");
         $data = [];
-        if (is_array($images)) {
-            foreach ($images as $value) {
+        foreach ($images as $value) {
                 $name = time().'_'.Str::random(10).'.'.$value->getClientOriginalExtension();
-                $value->storeAs('public/images/', $name);
+                $value->storeAs('public/images', $name);
                 $NewName = asset('/storage/images/'.$name);
-                $data[] = [
-                    $NewName
-                ];
-            }
-        }else {
-            $name = time().'_'.Str::random(10).'.'.$images->getClientOriginalExtension();
-            $images->storeAs('public/images', $name);
-            $NewName = asset('/storage/images/'.$name);
-                $data[] = [
-                    $NewName
-                ];
+                $data[] = $NewName;
         }
         return ResponseController::response($data);
     }
 
-    public function destroy($name)
+    public function destroy($fileName)
     {
+        $path = storage_path('app/public/images/'.$fileName);
 
+        if (!$path) {
+            return ResponseController::error('Image does not exist');
+        }
+        File::delete($path);
+        return ResponseController::success('Image has been successfully deleted');
     }
 }
